improve

+++ src/Outputs/FileOutput.php
@@ -9,13 +9,24 @@ class FileOutput extends \SplFileObject implements OutputInterface
 {
     use Concerns\InteractsWithOutput;
 
+    protected bool $autoDelete = true;
+
     public function __destruct()
     {
-        @\unlink($this->getRealPath());
+        if ($this->autoDelete) {
+            @\unlink($this->getRealPath());
+        }
     }
 
     public function getValue(): \SplFileObject
     {
+        return $this;
+    }
+
+    public function autoClean(bool $autoClean = true): static
+    {
+        $this->autoDelete = $autoClean;
+
         return $this;
     }
 }
